requirements.txt
pandas>=2.0
numpy>=1.24
paho-mqtt>=1.6.1

